==> 1_Car_price_prediction/notebooks/00_exploration.ipynb <==
DATA_PATH = "../data/raw/car_price.csv"

# import packages
import pandas as pd

pd.options.display.max_columns=100

## Reading data

raw_df = pd.read_csv(DATA_PATH)
raw_df.shape

# Dsiplay a random sample to build intuition about data
raw_df.sample(1).iloc[0]

# print general info. of the data frame
raw_df.info()

# Get some stats of the numerical columns
raw_df.describe()

# Lets take a look at the questionable columns
questionable_cols = ["Make", "Engine Fuel Type", "Engine Cylinders",]
for col in questionable_cols:
    print(col, "====> Uniques=", raw_df[col].nunique())
    print(raw_df[col].unique().tolist())
    print("----------------------------")

**We noticed that there are some cars that doesn't mention the number of Engine Cylinders so we have to deal with it**

pd.DataFrame(raw_df.groupby("Make")["Model"].nunique().sort_values(ascending=False))

# Check number of null values in each column
raw_df.isnull().sum()

# Check the number of each unique type of values in transmission type feature
raw_df[['Transmission Type']].value_counts()

# Number of duplicated rows:
print(raw_df.duplicated(keep='first').sum())
