==> 1_Car_price_prediction/notebooks/01_preprocessing.ipynb <==
DATA_PATH = "../data/raw/car_price.csv"
EXPORT_PATH = "../data/processed/00_preprocessed.pkl"

# load packages
import pandas as pd
import numpy as np

pd.options.display.max_columns=100

## Reading data

raw_df = pd.read_csv(DATA_PATH)
df = raw_df.copy()

## Modify columns names

# First we change the column names
df.columns = df.columns.str.lower().str.replace(" ", "_")
df.columns.tolist()

## replace the space in values with '_'

string_cols = list(df.dtypes[df.dtypes == "object"].index)
print(string_cols)
print("=============================")
for col in string_cols:
    df[col] = df[col].str.lower().str.replace(" ", "_")
    print(col, "====> Uniques=", df[col].nunique())
    print(df[col].unique()[:5])
    print("-------------------------")

## Quick sanity check:

print(raw_df.columns)
print(df.columns)

i = np.random.randint(0, raw_df.shape[0])
print("raw number", i, "\n------------------------------------------")
print(raw_df.iloc[i])
print("------------------------------------------")
print(df.iloc[i])

___
## Car Price Range:
**We can see that the MSRP (car price) is somewhat has a large range se we can scale it with a log exponentiation:**

## Export DataFrame into pickle file

df.to_pickle(EXPORT_PATH)
