==> 1_Car_price_prediction/notebooks/03_feature_engineering.ipynb <==
DATA_PATH = "../data/processed/00_preprocessed.pkl"
EXPORT_PATH = "../data/processed/01_OHE_data.pkl"

DROP_COLS = ["model", "engine_fuel_type", "driven_wheels",
             "number_of_doors", "market_category",
             "vehicle_size", "vehicle_style", "popularity"]

TRANSMISSION_DICT = {'manual': 1, 'automatic': 2,
                     'automated_manual': 3, 'direct_drive':4}



# Load packages
import pandas as pd
import numpy as np

from sklearn.preprocessing import OneHotEncoder

___
## Reading data:

df = pd.read_pickle(DATA_PATH)

df.head()

___
### Construct a list changer for the year columns:

lst = sorted(df.year.unique().tolist())
YEAR_DICT = {n: i+1 for i, n in enumerate(lst)}
YEAR_DICT

___
## Dropping irrelevant features:

df.drop(DROP_COLS, axis=1, inplace=True)
df.head()

___
### replace values for tranmsission type and year columns:

df.replace(TRANSMISSION_DICT, inplace=True)

df.replace(YEAR_DICT, inplace=True)

**Note we have some columns with 'unknown' value in the transmission type columns so we will need to remove it in cleaning phase**

___
## Transforming columns:
**Including the highway_mpg and city_mpg columns into one column that represents the mean speed of both and then applying minmaxscaler into it**

df['speed'] = (df['highway_mpg'] + df['city_mpg']) / 2

___
## OneHotEncoding the 'make' column:

# Building the onehotencoder
ohe = OneHotEncoder(handle_unknown='ignore')

# Fitting and transforming the data
ohe.fit(df[['make']])
transformed = ohe.transform(df[['make']])

# Merging the new columns into the dataframe
df[ohe.categories_[0]] = transformed.toarray()

transformed.toarray()[0]

print(ohe.categories_[0])

___
## Showing some sample from the dataframe after transforming and feature engineering:

df.sample(1).iloc[0]

___
## Now dropping tables that wont be used in our model:

DROPINGS = ["make", "highway_mpg", "city_mpg"]
df.drop(DROPINGS, axis=1, inplace=True)

df.sample(1).iloc[0]

___
## Export DataFrame:

df.to_pickle(EXPORT_PATH)
