==> 1_Car_price_prediction/notebooks/04b_baseline_model.ipynb <==
DATA_PATH = "../data/processed/02_cleaned_data.pkl"

MLFLOW_TRACKING_URI = "../models/mlruns"
MLFLOW_EXPERIMENT_NAME = "car_price_prediction"

LOG_PATH = "../models/temp/"
LOG_DATA_PKL = "data.pkl"
LOG_MODEL_PKL = "model.pkl"
LOG_METRICS_PKL = "metrics.pkl"

# Load packages
import pandas as pd

from pathlib import Path
import os
import pickle

from sklearn.preprocessing import RobustScaler, MinMaxScaler, StandardScaler
from sklearn.linear_model import LinearRegression
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import train_test_split

from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import r2_score

import mlflow
from mlflow.tracking import MlflowClient

pd.options.display.max_columns = 100

___
## Functions

def calculate_quality(ground_truth, predictions, metric_function, model_name):
    """
    Calculate the quality of the model according to different metric scores
    Input:
        ground_truth: from real observed data
        predictions: the predicted values from the model
        metric_function: the metric score funcrion used to measure performance
    Output:
        A dict of all scores for the given inputs
    """
    quality_score = {}
    quality_score[model_name] = round(metric_function(ground_truth, predictions), 3)
    quality_score = pd.Series(quality_score.values(), index=quality_score.keys())
    
    return quality_score

___
## Itialize
### Create directories:

Path(MLFLOW_TRACKING_URI).mkdir(parents=True, exist_ok=True)
Path(LOG_PATH).mkdir(parents=True, exist_ok=True)

___
## Read data:

df = pd.read_pickle(DATA_PATH)
df.sample(5)

___
## Splitting data into train and test:

X_train, X_test, y_train, y_test = train_test_split(df.drop('msrp', axis=1),
                                                    df[['msrp']],
                                                    test_size=0.2,
                                                    random_state=42)

# quick check of splitted data:
print("X_train shape", X_train.shape)
print("X_test shape", X_test.shape)
X_train.sample(5)

___ 
## Trying the base model:

### Intialize MLflow

# Intialize client and experiment
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

# Check if experiment already exists; if not create one
try:
    mlflow.create_experiment(MLFLOW_EXPERIMENT_NAME)
except:
    print(f'Experiment "{MLFLOW_EXPERIMENT_NAME}" exists at "{mlflow.get_tracking_uri()}"')

exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
print(exp)

**Now constructing the baseline model**
### Linear Regression:

lr = make_pipeline(MinMaxScaler(),
                   LinearRegression())

lr.fit(X_train.values, y_train.values)

**Evaluate on training data**

# Evaluate on training set:
predictions = pd.DataFrame(lr.predict(X_train.values),
                          columns=y_train.columns)

train_scores = {score.__name__:calculate_quality(y_train, predictions, score, "LinearRegression")
                for score in [r2_score, mean_squared_error, mean_absolute_error]}

train_scores = pd.concat(train_scores, axis=1)
train_scores

mean_train_score = train_scores.mean()
mean_train_score

**Evaluate on testing data**

# Evaluate on testing set:
predictions = pd.DataFrame(lr.predict(X_test.values),
                           columns=y_train.columns)

test_scores = {score.__name__:calculate_quality(y_test, predictions, score, "LinearRegression")
                for score in [r2_score, mean_squared_error, mean_absolute_error]}

test_scores = pd.concat(test_scores, axis=1)
test_scores

mean_test_score = test_scores.mean()
mean_test_score

___
## Log Run
**Now we save logs in mlflow directory**
### 1. Prepare

# Data details
data_details = {"data_path": DATA_PATH,
                "training_indices": X_train.index.tolist(),
                "test_indices": X_test.index.tolist(),
                "feature_names": X_train.columns.tolist(),
                "target_names": y_train.columns.tolist()}

with open(os.path.join(LOG_PATH, LOG_DATA_PKL), "wb") as output_file:
    pickle.dump(data_details, output_file)

# Model details
model = {"model_description": "Baseline model: LinearRegression",
         "model_details": str(lr),
         "model_object": lr}

with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), "wb") as output_file:
    pickle.dump(model, output_file)

# Performance details
regression_metrics = {"train_scores": train_scores,
                      "test_scores": test_scores}

with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), "wb") as output_file:
    pickle.dump(regression_metrics, output_file)

### 2. Logging

# Start a new run and track
with mlflow.start_run(experiment_id=exp.experiment_id, run_name=model["model_description"]):
    # log pickle
    mlflow.log_artifacts(LOG_PATH)
    
    # Track metrics
    for metric, score in mean_test_score.items():
        mlflow.log_metric(metric, score)
