==> 1_Car_price_prediction/notebooks/05_prediction_notebook.ipynb <==
MLFLOW_TRACKING_URI = "../models/mlruns"
MLFLOW_RUN_ID = "871e33a94e9a42819d93d28209cfca7e"

LOG_DATA_PKL = "data.pkl"
LOG_MODEL_PKL = "model.pkl"
LOG_METRIC_PKL = "metrics.pkl"


YEAR_DIC = {1990: 1, 1991: 2, 1992: 3, 1993: 4, 1994: 5, 
            1995: 6, 1996: 7, 1997: 8, 1998: 9, 1999: 10, 
            2000: 11, 2001: 12, 2002: 13, 2003: 14, 2004: 15,
            2005: 16, 2006: 17, 2007: 18, 2008: 19, 2009: 20,
            2010: 21, 2011: 22, 2012: 23, 2013: 24, 2014: 25, 
            2015: 26, 2016: 27, 2017: 28}

TRANSMISSION_DICT = {'manual': 1, 'automatic': 2,
                     'automated_manual': 3, 'direct_drive':4}

# Load packages
import pandas as pd
import numpy as np

import pickle
import os

import mlflow
from mlflow.tracking import MlflowClient

___
## Intilalize


# Intialize client
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

run = client.get_run(MLFLOW_RUN_ID)
artifacts_path = run.info.artifact_uri
artifacts_path

### Load artifacts

# Load model
model_path = os.path.join(artifacts_path, LOG_MODEL_PKL)
with open (model_path, "rb") as file:
    model = pickle.load(file)

model.keys()

# Load Data pickle
data_path = os.path.join(artifacts_path, LOG_DATA_PKL)
with open(data_path, "rb") as file:
    data = pickle.load(file)

data.keys()

# Unpack variables
features_names = pd.Series(data["feature_names"])
regressor = model["model_object"]

___
## Sample Entry

# [make, year, engine_hp, engine_cylinders, transmission_type, highway_mpg, city_mpg]
sample = ["mercedes-benz", 2014, 201, 4, 'manual', 34, 20]

sample_df = pd.DataFrame(features_names, columns=['features'])
sample_df['value'] = sample_df['features'].isin(sample)
sample_df

# Changing the values in sample dataframe
sample_df.iloc[[0],[1]] = YEAR_DIC[sample[1]]
sample_df.iloc[[1],[1]] = sample[2]
sample_df.iloc[[2],[1]] = sample[3]
sample_df.iloc[[3],[1]] = TRANSMISSION_DICT[sample[4]]
sample_df.iloc[[4],[1]] = (sample[5] + sample[6])/2

sample_df.replace({False:0, True:1}, inplace=True)

sample_df.set_index("features", inplace=True)

sample_df = sample_df.transpose()#.values
sample_df

___
## Predict

predictions = regressor.predict(sample_df.values)
predictions

car_price = round(np.expm1(predictions)[0], 3)
print(f"The estimated price of your car= ({car_price}) USD")
