==> 1_Car_price_prediction/notebooks/06a_test_functionality.ipynb <==
MLFLOW_TRACKING_URI = "../models/mlruns"
MLFLOW_RUN_ID = "871e33a94e9a42819d93d28209cfca7e"

import sys

sys.path.append("../scripts/")

from Car_Salary_Prediction import *

___
## Test functionality

model = PricePrediction(MLFLOW_TRACKING_URI, MLFLOW_RUN_ID)

sample = ["mercedes-benz", 2014, 201, 4, 'manual', 34, 20]

car_price = model.predict_car_price(sample)

car_price
