==> 1_Car_price_prediction/notebooks/06b_teat_api.ipynb <==
API_URL = "http://127.0.0.1:5000/"
PREDICTION_ENDPOINT = "/Predict_Car_Price"

# Load packages
import pandas as pd
import json
import requests

___
## Predict

sample = ["mercedes-benz", 2014, 201, 4, 'manual', 34, 20]

predictions = requests.post(API_URL + PREDICTION_ENDPOINT,
                            data=json.dumps(sample),
                            headers={"content-type": "application/json"})

predictions

predictions.json()

print(f"The estimated price of your car= ({predictions.json()}) USD")

___
## Another test sample

sample = ["bugatti", 2008, 1001, 16, 'automated_manual', 14, 8]
predictions = requests.post(API_URL + PREDICTION_ENDPOINT,
                            data=json.dumps(sample),
                            headers={"content-type": "application/json"})
print(f"The estimated price of your car= ({predictions.json()}) USD")

sample = ["bmw", 2011, 335, 6, 'manual', 26, 19]
predictions = requests.post(API_URL + PREDICTION_ENDPOINT,
                            data=json.dumps(sample),
                            headers={"content-type": "application/json"})
print(f"The estimated price of your car= ({predictions.json()}) USD")
